% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-values.R
\name{mi_column_values}
\alias{mi_column_values}
\title{Get column values for filtering a source}
\usage{
mi_column_values(source_name, year, level, filters = list(), limit = 40)
}
\arguments{
\item{source_name}{A \code{character} string specifying the source name (f_resource).}

\item{year}{A \code{character} or \code{integer} specifying the year.}

\item{level}{A \code{character} string specifying the NUTS level ("0", "1", "2", or "3").}

\item{filters}{A \verb{named list} where the names are the filter fields and
the values are the selected values for those fields. Default is an empty list.}

\item{limit}{an \code{integer} of length 1 specifying the maximum number of sources to return. Defaults to 1000.}
}
\value{
A \code{tibble} with the fields, labels, and their possible values for filtering.
}
\description{
Fetches the possible filtering values for a given source, year, and NUTS level.
}
\examples{
\dontrun{
mi_column_values(
  source_name = "DEMO_R_FIND2",
  year = 2020,
  level = "2",
  filters = list(unit = "YR")
)
}
}
