% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.R
\name{mi_nuts_levels}
\alias{mi_nuts_levels}
\title{Get a list of available NUTS levels}
\usage{
mi_nuts_levels()
}
\value{
a \code{character} vector of valid NUTS levels that will be accepted by other functions.
}
\description{
Get a list of available NUTS levels
}
\examples{
mi_nuts_levels()

}
