% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{mi_sources}
\alias{mi_sources}
\title{Get a list of available data sources}
\usage{
mi_sources(level, year = NULL, limit = 2500)
}
\arguments{
\item{level}{a \code{character} string specifying the NUTS level ("0", "1", "2", or "3"). You can also always check valid NUTS levels using \code{\link{mi_nuts_levels}}.}

\item{year}{an \code{integer} of length 1, specifying the year. Optional.}

\item{limit}{An \code{integer} specifying the maximum number of results to return. Default is 2500.}
}
\value{
a \code{tibble} of sources with the following columns:
\itemize{
\item \code{source_name}: name of the data source
\item \code{short_description}: short description of the data source
\item \code{description}: description of the data source
}
}
\description{
Get a list of available data sources
}
\examples{

\donttest{
# get up to 10 sources for NUTS level 3
mi_sources("3", limit = 10)

# get all sources for NUTS level 3 and year 2020
mi_sources("3", year = 2020)

}

}
