% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mi_data}
\alias{mi_data}
\title{Get univariate or bivariate data for a specific source}
\usage{
mi_data(
  x_source,
  y_source = NULL,
  year,
  level,
  x_filters = list(),
  y_filters = NULL,
  limit = 2500
)
}
\arguments{
\item{x_source}{A \code{character} string specifying the source name for the x variable.}

\item{y_source}{(Optional) A \code{character} string specifying the source name for the y variable.}

\item{year}{A \code{character} or \code{integer} specifying the year.}

\item{level}{A \code{character} string specifying the NUTS level ("0", "1", "2", or "3").}

\item{x_filters}{A \verb{named list} where the names are the filter fields for the x variable and the values are the selected values for those fields. Default is an empty list. To find out which filters to use, use \code{\link{mi_source_filters}} with the desired \code{source_name}.}

\item{y_filters}{(Optional) A \verb{named list} where the names are the filter fields for the y variable and the values are the selected values for those fields. Default is \code{NULL}. To find out which filters to use, use \code{\link{mi_source_filters}} with the desired \code{source_name}.}

\item{limit}{An \code{integer} specifying the maximum number of results to return. Default is 2500.}
}
\value{
A \code{tibble} with the following columns:
\itemize{
\item \code{geo}: code for the (NUTS) region at the requested level.
\item \code{geo_name}: name of the (NUTS) region at the requested level.
\item \code{geo_source}: source (type) of the spatial units at the requested level.
\item \code{geo_year}: year of the (NUTS) region at the requested level.
\item \code{x_year}: The year of the predictor variable (X), included in bivariate requests.
\item \code{y_year} (optional): The year of the outcome variable (Y), included in bivariate requests (only included when \code{y_source} is provided).
\item \code{x}: the value of the univariate variable.
\item \code{y} (optional): the value of the y variable (only included when \code{y_source} is provided).
}
}
\description{
Fetches univariate or bivariate data for a given source, year, NUTS level, and selected filters.
}
\examples{
\donttest{
# Univariate example
mi_data(
  x_source = "TGS00010",
  year = 2020,
  level = "2",
  x_filters = list(isced11 = "TOTAL", unit = "PC", age = "Y_GE15", freq = "A")
)

# Bivariate example
mi_data(
  x_source = "TGS00010",
  y_source = "DEMO_R_MLIFEXP",
  year = 2020,
  level = "2",
  x_filters = list(isced11 = "TOTAL", unit = "PC", age = "Y_GE15", freq = "A"),
  y_filters = list(unit = "YR", age = "Y_LT1", freq = "A")
)
}
}
