% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{mi_source_coverage}
\alias{mi_source_coverage}
\title{Get NUTS level and Year coverage for a specific source}
\usage{
mi_source_coverage(source_name)
}
\arguments{
\item{source_name}{name of the data source}
}
\value{
a \code{tibble} containing the following columns:
\itemize{
\item \code{nuts_level}: NUTS level
\item \code{year}: year
\item \code{source_name}: name of the data source (mathces the \code{source_name} requested by the user)
\item \code{short_description}: short description of the data source
\item \code{description}: description of the data source
}
}
\description{
Get the NUTS level and Year coverage for a specific data source.
}
\examples{
\dontrun{
mi_source_coverage("BD_HGNACE2_R3")

mi_source_coverage("ghs_smod")
}
}
